servicio de duplicado

+++ app/Http/Livewire/AdministracionCertificaciones.php
@@ -89,9 +89,13 @@ class AdministracionCertificaciones extends Component
     }
 
     public function delete(Certificacion $certificacion){
-            $certificacion->Hoja->update(['estado'=>3]);
+           if($certificacion->Hoja->update(['estado'=>3])){
             $certificacion->delete();
             $this->emitTo('administracion-certificaciones','render');
+           } else{
+            $this->emit("minAlert",["titulo"=>"AVISO DEL SISTEMA","mensaje"=>"Ocurrio un error al cambiar el estado de este certificado","icono"=>"warning"]);
+           }
+           
     }
